rename to elastic_app_search output plugin
