+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.0.0.beta1
-  - Changed documentation to correct required fields and other information [#2](https://github.com/logstash-plugins/logstash-output-appsearch/pull/2)
-  - Added check for correct host, engine and api_key during register [#2](https://github.com/logstash-plugins/logstash-output-appsearch/pull/2)
-  - Changed config `store_timestamp` to `timestamp_destination` [#2](https://github.com/logstash-plugins/logstash-output-appsearch/pull/2)
+  - Changed documentation to correct required fields and other information [#2](https://github.com/logstash-plugins/logstash-output-elastic_app_search/pull/2)
+  - Added check for correct host, engine and api_key during register [#2](https://github.com/logstash-plugins/logstash-output-elastic_app_search/pull/2)
+  - Changed config `store_timestamp` to `timestamp_destination` [#2](https://github.com/logstash-plugins/logstash-output-elastic_app_search/pull/2)
 
 ## 0.1.0
   - Initial version
