--- DEVELOPER.md
@@ -1,2 +1,0 @@
-# logstash-output-appsearch
-Example output plugin. This should help bootstrap your effort to write your own output plugin!
