+++ lib/logstash-output-elastic_app_search_jars.rb
